misc:runstat/profit: move bus profit from runstat to profit module
RunStat and MarketRunStat contained a lot of attributes which are not logically
clear to do statistics at runtime, and it makes hard to maintain runstat module.
As an enhancement to runstat, we keep runstat as a high level framework to do
statistics and separate each type of statistics into a new class and maintain it
independently in different modules.

Signed-off-by: Zhengwang Ruan <ruan.zhengwang@gmail.com>