core:emulation/futures: reimplement the emulation framework for Futures
Reimplement the emulation framework using multiprocessing to avoid Python GIL
and makes use of multi cores on hardware to improve the performance of parallel
computation. The major change is to start an independent process for each
contract instead of a thread, and introduce NSP and OSP mode to greatly reduces
the WP operations.

Signed-off-by: Zhengwang Ruan <ruan.zhengwang@gmail.com>